backend/requirements.txt
fastapi
uvicorn
temporalio
pydantic
python-dotenv
openai
pypdf
python-multipart
watchdog
sqlmodel
psycopg2-binary
passlib
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
requests